convert str to int

+++ application/controllers/Presence.php
@@ -215,7 +215,7 @@ class Presence extends Admin_Controller
 			for ($i=5; $i <= $rows; $i++) {
 				if ($i % 2 == 1) {
 					if (!empty($data->val($i, 3))) {
-						$record = preg_replace('/[\x00-\x1F\x7F]/u', '', $data->val($i, 3));
+						$record = (int) preg_replace('/[\x00-\x1F\x7F]/u', '', $data->val($i, 3));
 						if ($this->presence->is_employee_nonshift($record)) {
 							$employee[] = $record;
 						if ($tipe != 'auto') {
